requirement.txt
# Core dependencies
gradio==4.8.0
python-chess==1.999
openai==1.3.0
python-dotenv==1.0.0

# Optional for better performance
numpy==1.24.0
pandas==2.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0